config_loader.py
import yaml

def load_config(file_path):
    with open(file_path, 'r') as file:
        config = yaml.safe_load(file)
    return config

if __name__ == "__main__":
    config_data = load_config('config.yml')
    print(config_data)
