requirements.txt
streamlit
opencv-python
pillow
numpy
PyYAML
pika
pymongo
